Add two c6.99 mock configs for bootstrapping the i386 environment.
The initial build .00 is built purely against f19 to bootstrap the environment.
The second build .01 is built against the results of .00 which will have picked up a bunch of f19 dependencies which will fail now. We need to get rid of these.

The third build to be pushed in a later commit should be looking cleaner then and have most of these packages build fine and mostly the right dependencies.